+++ bookathlon-backend/src/main/java/com/bookathlon/service/LibroService.java
@@ -8,7 +8,7 @@ public interface LibroService {
 
 	
 	List<Libro> getLibri();
-	Libro getLibroById(int id);
+	Libro getLibroById(Long id);
 	Libro addLibro(Libro l);
 	List<Libro> getLibriDiTendenza();
